+++ sdk/lib/drivers/sound/mment4/detect.c
@@ -153,7 +153,7 @@ DetectNt4SoundDevices(
     IN  PWSTR BaseDeviceName,
     IN  SOUND_DEVICE_DETECTED_PROC SoundDeviceDetectedProc)
 {
-    ULONG DeviceNameLength = 0;
+    SIZE_T DeviceNameLength = 0;
     PWSTR DeviceName = NULL;
     ULONG Index = 0;
     HANDLE DeviceHandle;
