+++ win32ss/user/user32/windows/defwnd.c
@@ -869,7 +869,7 @@ RealDefWindowProcA(HWND hWnd,
         {
             PWSTR buf = NULL;
             PSTR outbuf = (PSTR)lParam;
-            UINT copy;
+            SIZE_T copy;
 
             if (Wnd != NULL && wParam != 0)
             {
